requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4
psycopg2-binary>=2.8
django-environ>=0.4.5
drf-yasg>=1.20.0
pytest>=6.2.4
pytest-django>=4.4.0
gunicorn>=20.1.0
